% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_acess_unitaria.R
\name{analisar_acess_unitaria}
\alias{analisar_acess_unitaria}
\title{Análise da acessibilidade unitária}
\usage{
analisar_acess_unitaria(munis = "todos", n_cores = 10)
}
\arguments{
\item{munis}{Uma string. A sigla do município, como utilizado no projeto -
uma sigla de três letras. Se \code{"todos"} (o padrão), roda a análise para
todos os municípios do projeto.}

\item{n_cores}{Um numérico. O número de cores para usar no cálculo da
acessibilidade. O padrão é 10 (metade dos cores do servidor).}
}
\value{
Uma lista com hexágonos que tenham acessibilidade unitária menor ou
igual a 15 e maior do que 0, pra cada cidade. Retornada de forma invisível.
}
\description{
Calcula a acessibilidade para cada cidade como se houvesse uma única
oportunidade em cada uma de suas células. Importante: lembrar de mudar o
parâmetro de memória a ser utilizada pelo java:
\code{options(java.parameters = "-Xmx64G")}.
}
\section{Detalhes}{

Calcula a acessibilidade usando a r5r::accessibility() como sem em cada
hexágono houvesse uma oportunidade. Importante notar que literalmente todas
as células das cidades contém uma oportunidade, e não apenas as que são
utilizadas no roteamento do projeto, de fato. Embora os mapas gerados nos
relatórios apresentem as células sem população e oportunidades de emprego,
saúde e educação como NA, isso é feito em uma etapa de pós-processamento,
apenas pra não apresentar como problemáticos hexágonos que não são
considerados no projeto. Para ilustrar porque isso é importante, vale olhar
os relatórios gerados pra cidade de São Paulo:
\itemize{
  \item{\code{relatorio_spo_full}: O resultado retornado pela
        r5r::accessibility(). Calcula a acessibilidade mesmo para célculas
        que não são utilizadas no projeto AOP porque não têm oportunidades e
        pessoas;}
  \item{\code{relatorio_spo}: O resultado como apresentado para todas as
        cidades. Mostra os resultados comentados no item acima, mas retrata
        as células sem oportunidades e pessoas como NA. Ou seja, os hexágonos
        problemáticos são hexágonos considerados na análise e que precisam
        ser tratados.}
  \item{\code{relatorio_spo_antigo}: o resultado que é obtido quando se
        calcula a acessibilidade apenas considerando os hexágonos
        considerados no projeto. Muitos hexágonos passam a serem considerados
        problemáticos não porque são pouco conectados com seus vizinhos, mas
        porque não têm nenhum vizinhos (ou têm poucos) considerados no
        projeto. O problema desses hexágonos, portanto, não está relacionado
        à rede viária, mas sim ao isolamento, que é uma informação que se
        obtém das análises do projeto. Esses hexágonos, portanto, não
        devem ser corrigidos.}
}
}

\examples{
if (interactive()) {

  options(java.parameters = "-Xmx64G")

  munis = c("rio")
  analisar_acess_unitaria(munis)

  # ou roda pra todos
  analisar_acess_unitaria()

}

}
