% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_conectividade_od.R
\name{analisar_conectividade_od}
\alias{analisar_conectividade_od}
\title{Análise da conectividade dos hexágonos como origens e destinos}
\usage{
analisar_conectividade_od(munis = "todos", n_cores = 10)
}
\arguments{
\item{munis}{Uma string. A sigla do município, como utilizado no projeto -
uma sigla de três letras. Se \code{"todos"} (o padrão), roda a análise para
todos os municípios do projeto.}

\item{n_cores}{Um numérico. O número de cores para usar no cálculo das
matrizes de tempo de viagem. O padrão é 10 (metade dos cores do servidor).}
}
\value{
Uma lista com hexágonos que alcancem menos que 15 destinos ou que
sejam alcançados por menos de 15 origens.
}
\description{
Calcula quantos destinos um hexágono pode alcançar, bem como quantas origens
conseguem alcançá-lo. Importante: lembrar de mudar o parâmetro de memória a
ser utilizada pelo java: \code{options(java.parameters = "-Xmx64G")}.
}
\section{Detalhes}{

A análise da acessibilidade levantou o questionamento de quantos hexágonos
poderiam estar isolados também no destino de um deslocamento, e não apenas em
sua origem. Essa análise mostra que todos os hexágonos isolados na origem
também o são na origem. A mesma "metodologia" utilizada na análise da
acessibilidade unitária, em que são calculadas as viagens também para
hexágonos que não são utilizados no projeto, se aplica nesta análise. Para
mais detalhes ver a documentação da função \code{analisar_acess_unitaria}.
}

\examples{
if (interactive()) {

  options(java.parameters = "-Xmx64G")

  munis = c("rio")
  analisar_conectividade_od(munis)

  # ou roda pra todos
  analisar_acess_unitaria()

}

}
