% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_resultados_geocode.R
\name{analisar_resultados_geocode}
\alias{analisar_resultados_geocode}
\alias{analisar_resultados_galileo}
\alias{gerar_analises_geocode}
\title{Análise dos resultados do geocode da nova base}
\usage{
analisar_resultados_geocode(base = c("address", "address_zip", "completo"))

analisar_resultados_galileo()

gerar_analises_geocode()
}
\arguments{
\item{base}{Qual das bases de análise utilizar. Utiliza o primeiro valor do
vetor.}
}
\value{
Chamada pelos side-effects, cria um gráfico com os resultados.
}
\description{
Calcula o percentual de erro do resultado do geocode feito com a nova base,
segundo as categorias apresentadas pelo Marcus: \code{VALID}, \code{GOOD}, \code{AVERAGE} e
\code{BAD}. Apresenta esses resultados também segundo o \code{score} de corte.

A função \code{analisar_resultados_galileo()} roda a mesma análise pros resultados
do Galileo (fora a análise por \code{score}, já que não existe esse campo), e a
\code{gerar_analises_geocode()} é uma wrapper pra rodar as análises pra todas as
bases.
}
