% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_snap.R
\name{analisar_snap}
\alias{analisar_snap}
\title{Análise do snap do R5}
\usage{
analisar_snap(munis = "todos")
}
\arguments{
\item{munis}{Uma string. A sigla do município, como utilizado no projeto -
uma sigla de três letras. Se \code{"todos"} (o padrão), roda a análise para
todos os municípios do projeto.}
}
\value{
Uma lista com a classificação dos snaps de cada cidade, em formato
  \code{c("data.table", "sf")}. Retornada de forma invisível.
}
\description{
Acha a distância do centroide de cada célula das grades utilizadas no projeto
até o ponto em que o snap do R5 de fato é feito. Importante: lembrar de mudar
o parâmetro de memória a ser utilizada pelo java:
\code{options(java.parameters = "-Xmx64G")}.
}
\section{Detalhes}{

Primeiro aplica a função \code{r5r::find_snap()} pra achar o snap dos
centroides das células das grades usadas no projeto.
Classifica esses snaps em quatro possíveis categorias:
\itemize{
  \item{Bom: dentro do hexágono, ou seja, a uma distância de até 174~ metros
        (tamanho da aresta do hexágono);}
  \item{Ok: dentro dos hexágonos imediatamente vizinhos, ou seja, a uma
        distãncia entre 174~ e 452~ metros (três vez a distância do centroide
        do hexágono ao ponto médio de sua aresta);}
  \item{Ruim: entre 452~ e 1600 metros; bem ruim, mas pelo menos achou algo.}
  \item{Péssimo: não conseguiu fazer o snap.}
}
Depois gera um relatório que inclui um mapa da distribuição de cada categoria
na cidade e uma tabela que resume o número de pessoas e oportunidades por
categoria. Salva esse relatório na pasta \code{resultados/analise_snap}.
Também salva a lista com a classificação do snap referente a cada célula das
grades, em formato \code{.rds}.
}

\examples{
if (interactive()) {

  options(java.parameters = "-Xmx64G")

  munis = c("rio")
  analisar_snap(munis)

  # ou roda pra todos
  analisar_snap()

}

}
