% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_snap.R
\name{apagar_arquivos_r5}
\alias{apagar_arquivos_r5}
\title{Apagar os arquivos gerados pelo R5}
\usage{
apagar_arquivos_r5(munis = "todos")
}
\arguments{
\item{munis}{Uma string. A sigla do município, como utilizado no projeto -
uma sigla de três letras. Se \code{"todos"} (o padrão), roda a análise para
todos os municípios do projeto.}
}
\value{
\code{NULL}, invisivelmente. Função chamada principalmente pelo seus
efeitos.
}
\description{
Apaga os arquivos gerados pelo R5, em formato \code{.mapdb}, \code{.mapdb.p}
e \code{.dat}.
}
\keyword{internal}
