% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_snap.R
\name{atualizar_pbfs}
\alias{atualizar_pbfs}
\title{Atualizar malhas viárias}
\usage{
atualizar_pbfs(munis = "todos")
}
\arguments{
\item{munis}{Uma string. A sigla do município, como utilizado no projeto -
uma sigla de três letras. Se \code{"todos"} (o padrão), roda a análise para
todos os municípios do projeto.}
}
\value{
\code{NULL}, invisivelmente. Função chamada principalmente pelo seus
efeitos.
}
\description{
Atualiza as malhas viárias em formato \code{.pbf} e apaga o grafo gerado pelo
R5.
}
\keyword{internal}
