% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_snap.R
\name{criar_relatorio}
\alias{criar_relatorio}
\title{Criar relatório de análise dos snaps}
\usage{
criar_relatorio(grade, sigla_muni, nome_muni)
}
\arguments{
\item{grade}{Um \code{c("data.table", "sf")} com a grade e dados do
município.}

\item{sigla_muni}{Uma string. A sigla do município.}

\item{nome_muni}{Uma string. O nome do município.}
}
\value{
Retorna o \code{data.table} passado no parâmetro \code{grade}
  invisivelmente.
}
\description{
Cria um relatório de análise dos snaps. Inclui um mapa com a geolocalização
dos snaps e uma tabela que vê a quantidade de pessoas e oportunidades por
categoria da classificação.
}
\keyword{internal}
