% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_acess_unitaria.R
\name{criar_relatorio_acess_unitaria}
\alias{criar_relatorio_acess_unitaria}
\title{Criar relatório de análise da acessibilidade unitária}
\usage{
criar_relatorio_acess_unitaria(grade, sigla_muni, nome_muni)
}
\arguments{
\item{grade}{Um \code{c("data.table", "sf")} com a grade e dados do
município.}

\item{sigla_muni}{Uma string. A sigla do município.}

\item{nome_muni}{Uma string. O nome do município.}
}
\value{
Retorna o \code{data.table} passado no parâmetro \code{grade}
invisivelmente.
}
\description{
Cria um relatório de análise da acessibilidade unitária. Inclui um mapa com a
distribuição da acessibilidade, destacando quais são os hexágonos
problemáticos (com menos do que 15 de acessibilidade e mais do que 0).
Também apresenta uma tabela com hexágonos com acessibilidade menor do que 15,
dessa vez incluindo os com acessibilidade igual a 0. Hexágonos com
acessibilidade igual a 0 são aqueles em que o r5 não consegue fazer snap.
}
\keyword{internal}
