% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_conectividade_od.R
\name{criar_relatorio_conectividade_od}
\alias{criar_relatorio_conectividade_od}
\title{Criar relatório de conectividade dos hexágonos como origens e destinos}
\usage{
criar_relatorio_conectividade_od(grade, sigla_muni, nome_muni)
}
\arguments{
\item{grade}{Um \code{c("data.table", "sf")} com a grade e dados do
município.}

\item{sigla_muni}{Uma string. A sigla do município.}

\item{nome_muni}{Uma string. O nome do município.}
}
\value{
Retorna o \code{data.table} passado no parâmetro \code{grade}
invisivelmente.
}
\description{
Cria um relatório de conectividade dos hexágonos como origens e destinos.
Inclui mapas com a distribuição do número de destinos alcançados por hexágono
e com a distribuição do números de origens que alcançam cada hexágono, bem
como tabelas que ajudam a identificar hexágonos problemáticos.
}
\keyword{internal}
